requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
celery-redbeat==2.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.5
django-contrib-comments==2.1.0
django-debreach==2.0.1
django-import-export==2.5.0
django-js-reverse==0.9.1
django-log-request-id==1.6.0
django-model-utils==4.1.1
django-webpack-loader==1.1.0
djangorestframework==3.12.4
docopt==0.6.2
docx==0.2.4
docx2pdf==0.1.7
docxtpl==0.11.5
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
img2pdf==0.4.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
odfpy==1.4.1
openpyxl==3.0.7
parso==0.8.2
pep517==0.11.0
pickleshare==0.7.5
pikepdf==2.16.1
Pillow==8.3.1
pip-tools==6.2.0
pipreqs==0.4.10
prompt-toolkit==3.0.19
psutil==5.8.0
psycopg2==2.9.1
pycparser==2.20
Pygments==2.9.0
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
pytz==2021.1
pywin32==301; platform_system == "Windows"
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
sentry-sdk==1.3.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tenacity==8.0.1
tomli==1.0.4
tqdm==4.61.2
traitlets==5.0.5
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
dropbox
tika
